Rename /test --> /integration_test
- move directory
- remove 'krew' subpackage (merge with main package)
- make the package _test.go-only, by merging /test/krew package
- fix a bug with run-integration-tests.sh where it silently failed when the
  binary file didn't exist (due to readlink failing inside the subshell).
- use testutil.TempDir() in the integration tests (per validate-code-patterns)
